requirements.txt
pandas 
scikit-learn 
numpy 
redis
elasticsearch 
sqlalchemy
kafka-python